Добавил фильтр на импорт номенклатур
