+++ ValidationRules.StateInitialization.Host/DataObjectTypesProviderFactory.cs
@@ -77,7 +77,7 @@ namespace NuClear.ValidationRules.StateInitialization.Host
                             typeof(FirmFacts::Order),
                             typeof(FirmFacts::OrderPosition),
                             typeof(FirmFacts::OrderPositionAdvertisement),
-                            typeof(FirmFacts::Position),
+                            typeof(FirmFacts::SpecialPosition),
                             typeof(FirmFacts::Project),
                         });
             }
