+++ ValidationRules.Storage/Model/FirmRules/Facts/SpecialPosition.cs
@@ -1,9 +1,7 @@
 namespace NuClear.ValidationRules.Storage.Model.FirmRules.Facts
 {
-    public class Position
+    public class SpecialPosition
     {
         public long Id { get; set; }
-        public long CategoryCode { get; set; }
-        public int Platform { get; set; }
     }
 }
