+++ ValidationRules.Storage/ValidationRules.Storage.csproj
@@ -234,7 +234,7 @@
     <Compile Include="Model\FirmRules\Facts\Order.cs" />
     <Compile Include="Model\FirmRules\Facts\OrderPosition.cs" />
     <Compile Include="Model\FirmRules\Facts\OrderPositionAdvertisement.cs" />
-    <Compile Include="Model\FirmRules\Facts\Position.cs" />
+    <Compile Include="Model\FirmRules\Facts\SpecialPosition.cs" />
     <Compile Include="Model\FirmRules\Facts\Project.cs" />
     <Compile Include="Model\PriceRules\Aggregates\AdvertisementAmountRestriction.cs" />
     <Compile Include="Model\PriceRules\Aggregates\AmountControlledPosition.cs" />
